+++ assets/wizards/popups/views/segments/index.js
@@ -2,8 +2,8 @@
  * Internal dependencies.
  */
 import { withWizardScreen } from '../../../../components/src';
-import SegmentsList from './SegmentsList';
-import SingleSegment from './SingleSegment';
+import SegmentsList from './segments-list';
+import SingleSegment from './single-segment';
 import './style.scss';
 
 /**