+++ assets/wizards/popups/views/segments/single-segment.js
@@ -19,7 +19,7 @@ import {
 	TextControl,
 	hooks,
 } from '../../../../components/src';
-import SubscriptionListsControl from '../../../../components/src/subscription-lists-control';
+import ListsControl from '../../components/lists-control';
 
 const { useHistory } = Router;
 const { SettingsCard, SettingsSection, MinMaxSetting } = Settings;
@@ -95,6 +95,7 @@ const SingleSegment = ( { segmentId, setSegments, wizardApiFetch } ) => {
 				criteria: segmentCriteria,
 				configuration: segmentConfig,
 			},
+			quiet: true,
 		} )
 			.then( setSegments )
 			.then( history.push( '/segments' ) );
@@ -112,7 +113,7 @@ const SingleSegment = ( { segmentId, setSegments, wizardApiFetch } ) => {
 		const config = [ ...segmentCriteria ];
 		const item = config.find( c => c.criteria_id === id );
 		if ( item ) {
-			if ( ! value ) {
+			if ( ! value || ( Array.isArray( value ) && 0 === value.length ) ) {
 				config.splice( config.indexOf( item ), 1 );
 			} else if ( ! Array.isArray( value ) && typeof value === 'object' ) {
 				item.value = { ...item.value, ...value };
@@ -192,7 +193,7 @@ const SingleSegment = ( { segmentId, setSegments, wizardApiFetch } ) => {
 
 			<SettingsCard
 				title={ __( 'Reader Engagement', 'newspack' ) }
-				description={ __( 'Target readers based on their browsing behavior', 'newspack' ) }
+				description={ __( 'Target readers based on their browsing behavior.', 'newspack' ) }
 				noBorder
 			>
 				{ allCriteria
@@ -208,8 +209,11 @@ const SingleSegment = ( { segmentId, setSegments, wizardApiFetch } ) => {
 					) ) }
 			</SettingsCard>
 			<SettingsCard
-				title={ __( 'Reader Activity', 'newspack' ) }
-				description={ __( 'Target readers based on their actions', 'newspack' ) }
+				title={ __( 'Registration', 'newspack' ) }
+				description={ __(
+					'Target readers based on their user account registration status.',
+					'newspack'
+				) }
 				columns={ 3 }
 				noBorder
 			>
@@ -226,8 +230,47 @@ const SingleSegment = ( { segmentId, setSegments, wizardApiFetch } ) => {
 					) ) }
 			</SettingsCard>
 			<SettingsCard
+				title={ __( 'Newsletters', 'newspack' ) }
+				description={ __(
+					'Target readers based on their newsletter subscription status.',
+					'newspack'
+				) }
+				columns={ 3 }
+				noBorder
+			>
+				{ allCriteria
+					.filter( criteria => criteria.category === 'newsletter' )
+					.map( criteria => (
+						<SettingsSection
+							key={ criteria.id }
+							title={ criteria.name }
+							description={ criteria.description }
+						>
+							{ getCriteriaInput( criteria ) }
+						</SettingsSection>
+					) ) }
+			</SettingsCard>
+			<SettingsCard
+				title={ __( 'Reader Revenue', 'newspack' ) }
+				description={ __( 'Target readers based on their revenue activity.', 'newspack' ) }
+				columns={ 3 }
+				noBorder
+			>
+				{ allCriteria
+					.filter( criteria => criteria.category === 'reader_revenue' )
+					.map( criteria => (
+						<SettingsSection
+							key={ criteria.id }
+							title={ criteria.name }
+							description={ criteria.description }
+						>
+							{ getCriteriaInput( criteria ) }
+						</SettingsSection>
+					) ) }
+			</SettingsCard>
+			<SettingsCard
 				title={ __( 'Referrer Sources', 'newspack' ) }
-				description={ __( 'Target readers based on where they’re coming from', 'newspack' ) }
+				description={ __( 'Target readers based on where they’re coming from.', 'newspack' ) }
 				notification={ __(
 					'Segments using these options will apply only to the first page visited after coming from an external source.',
 					'newspack'
@@ -295,8 +338,51 @@ addFilter(
 	function ( element, criteria, value, update ) {
 		if ( [ 'subscribed_lists', 'not_subscribed_lists' ].includes( criteria.id ) ) {
 			return (
-				<SubscriptionListsControl
+				<ListsControl
 					placeholder={ __( 'Start typing to search for lists…', 'newspack-plugin' ) }
+					path="/newspack-newsletters/v1/lists_config"
+					value={ value }
+					onChange={ update }
+				/>
+			);
+		}
+		return element;
+	}
+);
+
+/**
+ * Adds a custom input for the active_subscriptions and not_active_subscriptions criteria.
+ */
+addFilter(
+	'newspack.criteria.input',
+	'newspack.activeSubscriptions',
+	function ( element, criteria, value, update ) {
+		if ( [ 'active_subscriptions', 'not_active_subscriptions' ].includes( criteria.id ) ) {
+			return (
+				<ListsControl
+					placeholder={ __( 'Start typing to search for products…', 'newspack-plugin' ) }
+					path="/newspack/v1/wizard/newspack-popups-wizard/subscription-products"
+					value={ value }
+					onChange={ update }
+				/>
+			);
+		}
+		return element;
+	}
+);
+
+/**
+ * Adds a custom input for the active_memberships and not_active_memberships criteria.
+ */
+addFilter(
+	'newspack.criteria.input',
+	'newspack.activeMemberships',
+	function ( element, criteria, value, update ) {
+		if ( [ 'active_memberships', 'not_active_memberships' ].includes( criteria.id ) ) {
+			return (
+				<ListsControl
+					placeholder={ __( 'Start typing to search for membership plans…', 'newspack-plugin' ) }
+					path="/wc/v3/memberships/plans?per_page=100"
 					value={ value }
 					onChange={ update }
 				/>