+++ includes/class-recaptcha.php
@@ -14,7 +14,6 @@ defined( 'ABSPATH' ) || exit;
  */
 final class Recaptcha {
 	const SCRIPT_HANDLE  = 'newspack-recaptcha';
-	const THRESHOLD      = 0.5;
 	const OPTIONS_PREFIX = 'newspack_recaptcha_';
 
 	/**
@@ -124,6 +123,7 @@ final class Recaptcha {
 			'use_captcha' => false,
 			'site_key'    => '',
 			'site_secret' => '',
+			'threshold'   => 0.5,
 		];
 	}
 
@@ -288,7 +288,7 @@ final class Recaptcha {
 		// If the reCaptcha verification score is below our threshold for valid user input.
 		if (
 			isset( $captcha_verify['score'] ) &&
-			self::THRESHOLD > floatval( $captcha_verify['score'] )
+			floatval( self::get_setting( 'threshold' ) ) > floatval( $captcha_verify['score'] )
 		) {
 			return new \WP_Error(
 				'newspack_recaptcha_failure',
@@ -355,7 +355,9 @@ final class Recaptcha {
 	 * Verify reCAPTCHA v3 on checkout submission.
 	 */
 	public static function verify_recaptcha_on_checkout() {
-		if ( ! self::can_use_captcha() ) {
+		$url                   = \home_url( \add_query_arg( null, null ) );
+		$should_verify_captcha = apply_filters( 'newspack_recaptcha_verify_captcha', self::can_use_captcha(), $url );
+		if ( ! $should_verify_captcha ) {
 			return;
 		}
 		$token = isset( $_POST['g-recaptcha-response'] ) ? sanitize_text_field( wp_unslash( $_POST['g-recaptcha-response'] ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification.Missing