+++ includes/wizards/class-popups-wizard.php
@@ -8,6 +8,7 @@
 namespace Newspack;
 
 use \WP_Error, \WP_Query;
+use \Newspack\Donations;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -38,6 +39,7 @@ class Popups_Wizard extends Wizard {
 	public function __construct() {
 		parent::__construct();
 		add_action( 'rest_api_init', [ $this, 'register_api_endpoints' ] );
+		add_filter( 'newspack_popups_registered_criteria', [ $this, 'maybe_unregister_memberships_criteria' ] );
 	}
 
 	/**
@@ -371,6 +373,21 @@ class Popups_Wizard extends Wizard {
 						'sanitize_callback' => 'absint',
 					],
 					'name' => [
+						'sanitize_callback' => 'sanitize_text_field',
+					],
+				],
+			]
+		);
+		
+		register_rest_route(
+			NEWSPACK_API_NAMESPACE,
+			'/wizard/' . $this->slug . '/subscription-products',
+			[
+				'methods'             => \WP_REST_Server::READABLE,
+				'callback'            => [ $this, 'api_get_subscription_products' ],
+				'permission_callback' => [ $this, 'api_permissions_check' ],
+				'args'                => [
+					's' => [
 						'sanitize_callback' => 'sanitize_text_field',
 					],
 				],
@@ -916,5 +933,71 @@ class Popups_Wizard extends Wizard {
 		$cm = Configuration_Managers::configuration_manager_class_for_plugin_slug( 'newspack-popups' );
 		$cm->rename_campaign( $request['id'], $request['name'] );
 		return $this->api_get_settings();
+	}
+
+	/**
+	 * Get non-donation subscription products.
+	 * 
+	 * @param WP_REST_Request $request Full details about the request.
+	 * @return WP_REST_Response|WP_Error Response object on success, or WP_Error object on failure.
+	 */
+	public function api_get_subscription_products( $request ) {
+		$args = [
+			'post_type'      => 'product',
+			'posts_per_page' => 100,
+			'post_status'    => 'publish',
+			'tax_query'      => [ // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_tax_query
+				[
+					'taxonomy' => 'product_type',
+					'field'    => 'slug',
+					'terms'    => [ 'subscription', 'variable-subscription' ],
+				],
+			],
+		];
+
+		$posts = array_values(
+			array_filter(
+				\get_posts( $args ),
+				function( $post ) {
+					return ! Donations::is_donation_product( $post->ID );
+				}
+			)
+		);
+
+		return \rest_ensure_response(
+			array_map(
+				function( $post ) {
+					return [
+						'id'    => $post->ID,
+						'title' => $post->post_title,
+					];
+				},
+				$posts
+			)
+		);
+	}
+
+	/**
+	 * We only want to show Memberships criteria if the WooCommerce Memberships extension is active.
+	 * Otherwise these criteria are meaningless.
+	 * 
+	 * @param array $criteria Registered criteria.
+	 * 
+	 * @return array Filtered criteria.
+	 */
+	public function maybe_unregister_memberships_criteria( $criteria ) {
+		if ( ! class_exists( 'WC_Memberships' ) ) {
+			$memberships_criteria = [ 'active_memberships', 'not_active_memberships' ];
+			$criteria             = array_values(
+				array_filter(
+					$criteria,
+					function( $config ) use ( $memberships_criteria ) {
+						return ! in_array( $config['id'], $memberships_criteria, true );
+					}
+				)
+			);
+		}
+
+		return $criteria;
 	}
 }